fix: rename & cleanup
