+++ frontend/src/components/annotation/utils/AnnotationSavePrechecks.ts
